Reapproached

+++ src/main/java/com/adetola/guestbook/repository/UserRepository.java
@@ -3,6 +3,6 @@ package com.adetola.guestbook.repository;
 import com.adetola.guestbook.entity.GuestbookUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<GuestbookUser, Integer> {
+public interface UserRepository extends JpaRepository<GuestbookUser, Long> {
     GuestbookUser findByEmail(String email);
 }
